+++ src/HotChocolate/Core/src/Execution/Processing/IExecutionContext.cs
@@ -45,23 +45,9 @@ namespace HotChocolate.Execution.Processing
         ObjectPool<PureResolverTask> PureResolverTasks { get; }
 
         /// <summary>
-        /// Gets the batch task pool.
-        /// </summary>
-        ObjectPool<BatchExecutionTask> BatchTasks { get; }
-
-        /// <summary>
         /// Gets the task buffer pool.
         /// </summary>
         ObjectPool<IExecutionTask?[]> TaskBuffers { get; }
-
-        /// <summary>
-        /// Defines if a execution task can be processed of if it is suspended/paused.
-        /// </summary>
-        /// <param name="executionTask">
-        /// The execution task that shall be evaluated.
-        /// </param>
-        /// <returns></returns>
-        bool IsSuspended(IExecutionTask executionTask);
 
         /// <summary>
         /// Resets the execution state.
